adding query parameters to get requests

+++ upcoming/upcoming_data/src/main/java/com/rappi/upcoming_data/remote/dto/UpcomingDto.kt
@@ -4,7 +4,7 @@ import com.squareup.moshi.Json
 
 data class UpcomingDto (
     val dates: Dates?,
-    val page: Long?,
+    val page: Int?,
     val results: List<Upcoming?>?,
     @Json(name="total_pages")
     val totalPages: Long?,
